pyproject.toml
[tool.poetry]
name = "weather"
version = "0.1.0"
description = "Weather Data System"
authors = ["Edgaras Gacionis"]
license = "Turing"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.2"
sqlalchemy = "^2.0.30"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
black = "^24.4.2"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
